+++ CHANGELOG.md
@@ -145,6 +145,19 @@
     - `list` returns a list of payments
     - `delete` deletes a payment
     - `getFields` returns the fields and settings for payments
+- Added service `Services\Sale\CashboxHandler\Service\CashboxHandler` with support methods,
+  see [sale.cashbox.handler.* methods](https://github.com/bitrix24/b24phpsdk/issues/258):
+    - `add` adds a REST cashbox handler
+    - `update` updates the data of the REST cashbox handler
+    - `list` returns a list of available REST cashbox handlers
+    - `delete` deletes the REST cashbox handler
+- Added service `Services\Sale\Cashbox\Service\Cashbox` with support methods,
+  see [sale.cashbox.* methods](https://github.com/bitrix24/b24phpsdk/issues/258):
+    - `add` adds a new cash register
+    - `update` updates an existing cash register
+    - `list` returns a list of configured cash registers
+    - `delete` deletes a cash register
+    - `checkApply` saves the result of printing the receipt
 - Added service `Services\Calendar\Service\Calendar` with support methods,
   see [calendar.* methods](https://github.com/bitrix24/b24phpsdk/issues/263):
     - `add` adds a new calendar section